reuse existing configConfig
