+++ x/smart/strategy_cache.go
